fix: resolve all build errors and warnings
Build error fixes:
- usb_core.c: remove calls to unimplemented xhci_enable_slot,
  xhci_address_device, xhci_find_keyboard, xhci_keyboard_poll.
  Enumeration now uses only existing xhci_port_get_status/reset
  and records attached state. Full enumeration deferred to when
  XHCI slot/address support is implemented.

Warning fixes:
- auth_policy.h: lockout_duration_ticks uint32_t -> uint64_t (overflow)
- syscall.c: explicit (int64_t) casts on ternary returns (signedness)
- task.c: remove unused task_entry_wrapper declaration
- gpu_core.c: remove unused subclass variable
- ed25519.c: mark scaffolding field functions __attribute__((unused))
- boot_slot.c: mark bs_crc32 __attribute__((unused))
- context_switch.S: add .note.GNU-stack section (executable stack warning)

+++ include/core/auth_policy.h
@@ -19,7 +19,7 @@ struct auth_attempt {
 
 struct auth_policy_config {
   uint32_t max_attempts;
-  uint32_t lockout_duration_ticks;
+  uint64_t lockout_duration_ticks;
   int audit_enabled;
 };
